refactor: break up FetchSourceHandler into focused services (#43)
Extract ArticleEnrichmentService (categorize + summarize + keywords +
translate + score) and AlertDispatchService (match + dispatch) from
FetchSourceHandler, reducing it from 322 lines / 17 deps to 135 lines / 10 deps.
